No concept drift. Training on a standard Hoeffding tree.
Each step is 1000 instances, each episode is 20 steps. Each episode is a new stream.

Reward is (accuracy with agent) - (accuracy without agent) for a given step.

There are 25 states, indexed from 0 to 24 (rows). There are 5 actions, 0 to 4 (columns).

Agent can multiply or divide delta by a range of factors. This
determines how difficult or easy it is for Hoeffding Tree to split.

Each state is the current epoch's accuracy relative to last epochs accuracy and relative to the last 5 epochs accuracy.

These variables are each binned from 0 to 4 and the combinations are indexed from 0 to 24.