Design issue to solve: Utility of components for now is what their performance was on the problem they were trained on, not a general measure. Perhaps we can update utility as some function of the reward received for using a component for future problems.
We want to see if reuse and deletion alone automatically allows for creating an agent that can respond to arbitrary problems, but the utility of a tested component is linked to a previous problem that might be totally different. How should utility be updated and distributed in a simple manner without becoming a confounding factor in results that support or contradict the central thesis?